requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
asgiref==3.2.10
click==7.1.2
Django==3.1.1
djangorestframework==3.11.1
drf-nested-routers==0.91
gunicorn==20.0.4
importlib-metadata==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
zipp==3.2.0